Parse parse
The cookie we're receiving from the server is a json stringified
string. The first parse converts the cookie to a json string, the
second parse converts the json string into an object.

Yes, this is silly; no, we can't change it (unless we want to do
some reworking on the platform as well).