refactor: change runAdapter.query to better support multiplayer
Needed a way to support world-scoped runs in a multiplayer setting
where the worlds are scoped to an episode. The 'scope' argument
has been changed to be optional. When not a scope is not provided,
runAdapter.query will now default to using:

'run/in/groupName/[episodeName]'

Which allows you to query runs by episode even though the runs
themselves might not be scoped by them.

BREAKING CHANGE: query(modelFile, scope, optionals) -> query(modelFile, optionals)