refactor: episodeAdapter.byName -> episodeAdapter.withName (new args)
Would like to use the 'by' prefix strictly for functions that will
return a list instead of a single item. Switching to 'with' instead.

Also to ensure consistency with the rest of libs, moving groupName from
a required argument to an optional one, where the libs uses the groupName
from the session when it is not explicitly provided.

BREAKING CHANGE: episodeAdapter.byName -> withName; takes groupName as an optional argument