refactor: errorManager handlers to handle follow-up errors better
In particular, thinking of the case a GET for a resource results
in a 401, which can be handled generically by logging in, but on
retry a different error could be thrown, maybe a 404 if the
resource doesn't exist.

errorManager should now switch to trying to handle the error
instead of passing around the original one trying to fix it.