feat: config.setContext to configure protocol
This method serves as an alternate way to set the configuration
settings. It's somewhat unconventional to directly set the
property of class instance, so here's a function for that instead.

Main reasure for this is that the proxy (v3 node server) runs in
a sandbox that disallows something like "config.apiHost = 'forio.com'"